--- src/init.js
@@ -1,3 +1,0 @@
-import { registerDavProperty } from '@nextcloud/files'
-
-registerDavProperty('nc:approval-state', { nc: 'http://nextcloud.org/ns' })
